Add files via upload
A fix for the Fourth Dimensional Sight skill from Ashes of the Domain: Virtues of Society crashing the game when run with IX Revival. The 4DS skill is depreciated content in VoS so this fix will not make any gameplay changes. To install, unzip the data folder into the VoS folder and confirm overwriting the file.